Clean up to separate BrokerInfoSelector and BrokerSelector
