+++ pinot-clients/pinot-java-client/src/main/java/org/apache/pinot/client/SimpleBrokerSelector.java
@@ -16,11 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.broker;
+package org.apache.pinot.client;
 
 import com.google.common.collect.ImmutableList;
 import java.util.List;
 import java.util.Random;
+import org.apache.pinot.common.broker.BrokerSelector;
 
 
 /**
